refactor: migrate configuration from `config` module to `core` options - Move configuration structs related to filtering, HTML options, input source, listing, and more to the `core` layer for better modularity and reuse. - Streamline imports across modules to reduce dependencies on the higher-level configuration layer. - Ensure minimal changes to functionality while improving organization of the codebase.
