+++ docs/src/configuration_files.md
@@ -39,10 +39,13 @@ calculate_line_count = true
 format = "markdown"
 
 [llm]
-provider    = "openai"
-model       = "gpt-4o"
-api_key_env = "OPENAI_API_KEY"  # recommended – keeps secrets out of the file
-# api_key_file = "./.secrets/openai.key"   # ← if used, rustree checks file perms
+provider    = "openai"        # Supported: openai, anthropic, cohere, openrouter
+model       = "gpt-4o"         # Optional - uses provider's default if not specified
+api_key_env = "OPENAI_API_KEY" # Recommended – keeps secrets out of the file
+# api_key     = "sk-..."       # Direct API key (not recommended in shared configs)
+# endpoint    = "https://..."  # Custom endpoint for self-hosted or proxy services
+# temperature = 0.7            # Model temperature (0.0-2.0)
+# max_tokens  = 1000           # Maximum response tokens
 ```
 
 Unknown keys are ignored but a warning is printed, so your config keeps working
