+++ src/config/html.rs
@@ -1,45 +1,1 @@
-// src/config/html.rs
-
-//! Configuration specific to HTML output.
-//!
-//! These options are only consulted when `--output-format html` (or
-//! `LibOutputFormat::Html`) is active.  They correspond closely to the legacy
-//! flags provided by GNU `tree` (-H, --nolinks, --hintro, --houtro).
-
-use std::path::PathBuf;
-
-#[derive(Debug, Clone)]
-pub struct HtmlOptions {
-    /// If present, this string is prepended to every hyperlink that is
-    /// generated (e.g. "https://example.org/").  It should NOT contain a
-    /// trailing slash; the formatter takes care of inserting one.
-    pub base_href: Option<String>,
-
-    /// When `true`, the first path component of the generated relative link is
-    /// stripped.  This mimics the behaviour of GNU tree's "-H -baseHREF".
-    pub strip_first_component: bool,
-
-    /// Use a custom intro fragment (HTML that comes *before* the `<pre>` tree)
-    /// instead of the built-in default.
-    pub custom_intro: Option<PathBuf>,
-
-    /// Use a custom outro fragment (HTML that comes *after* the `<pre>` tree)
-    /// instead of the built-in default.
-    pub custom_outro: Option<PathBuf>,
-
-    /// Whether to generate `<a href>` hyperlinks.  If `false`, only plain text
-    /// (escaped) file names are shown.
-    pub include_links: bool,
-}
-
-impl Default for HtmlOptions {
-    fn default() -> Self {
-        Self {
-            base_href: None,
-            strip_first_component: false,
-            custom_intro: None,
-            custom_outro: None,
-            include_links: true,
-        }
-    }
-}
+pub use crate::core::options::html::*;
