+++ src/config/listing.rs
@@ -1,12 +1,1 @@
-/// Configuration for directory listing behavior.
-#[derive(Debug, Clone, Default)]
-pub struct ListingOptions {
-    /// Maximum depth to recurse into subdirectories. None means unlimited.
-    pub max_depth: Option<usize>,
-    /// Whether to show hidden files and directories (those starting with '.').
-    pub show_hidden: bool,
-    /// Whether to list only directories, excluding files.
-    pub list_directories_only: bool,
-    /// Whether to show the full relative path for each file/directory.
-    pub show_full_path: bool,
-}
+pub use crate::core::options::listing::*;
