+++ src/core/formatter/base.rs
@@ -1,6 +1,6 @@
 // src/core/formatter/base.rs
-use crate::config::RustreeLibConfig;
 use crate::core::error::RustreeError;
+use crate::core::options::RustreeLibConfig;
 use crate::core::tree::node::NodeInfo;
 
 /// A trait for formatting a list of `NodeInfo` objects into a string representation.
