+++ src/core/formatter/text_tree.rs
@@ -1,8 +1,8 @@
 use super::base::TreeFormatter;
-use crate::config::RustreeLibConfig;
 use crate::core::error::RustreeError;
 use crate::core::metadata::MetadataAggregator;
 use crate::core::metadata::file_info::{MetadataStyle, format_node_metadata};
+use crate::core::options::RustreeLibConfig;
 use crate::core::tree::node::{NodeInfo, NodeType};
 use std::collections::HashMap;
 use std::fmt::Write;
