+++ src/core/mod.rs
@@ -6,6 +6,7 @@ pub mod formatter;
 pub mod input;
 pub mod llm;
 pub mod metadata;
+pub mod options;
 pub mod sorter;
 pub mod tree;
 pub mod util;
