+++ src/core/options/input_source.rs
@@ -1,0 +1,12 @@
+/// Options related to the *source* of the directory tree that the library
+/// operates on.
+#[derive(Debug, Clone, Default)]
+pub struct InputSourceOptions {
+    /// Display name for the root directory/file.
+    pub root_display_name: String,
+    /// Size of the root node, if available and size reporting is enabled.
+    pub root_node_size: Option<u64>,
+    /// Whether the root path represents a directory (`true`) or a single file
+    /// (`false`).
+    pub root_is_directory: bool,
+}
